+++ src/Core/Console/ConsoleVisual.h
@@ -11,7 +11,7 @@ namespace oakvr
 		ConsoleVisual();
 
 		auto DrawHistory(std::vector<std::string> vecHistory) -> void;
-		auto DrawCommandline(std::string commandLine, int cursorPosition)->void;
+		auto DrawCommandline(std::string commandLine, size_t cursorPosition)->void;
 	private:
 		oakvr::render::Sprite m_historyBackground;
 		oakvr::render::Sprite m_cmdLineBackground;
